Remove all vtunnel refrences
All vtunnel refrences is removed out of the code,
the guest agent now only supports the namespaced
network.
