+++ src/go/guestagent/pkg/forwarder/vtunnel.go
@@ -1,5 +1,5 @@
 /*
-Copyright © 2022 SUSE LLC
+Copyright © 2024 SUSE LLC
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
@@ -12,47 +12,17 @@ limitations under the License.
 */
 
 // Package forwarder implements a forwarding mechanism to forward
-// port mappings over Vtunnel.
+// port mappings over the network.
 package forwarder
 
 import (
-	"encoding/json"
-	"net"
-
 	"github.com/rancher-sandbox/rancher-desktop/src/go/guestagent/pkg/types"
 )
 
 // Forwarder is the interface that wraps the Send method which
 // to forward the port mappings.
 type Forwarder interface {
-	// Send sends the give port mappings to the VTunnel Peer via
+	// Send sends the give port mappings to the Peer via
 	// a tcp connection.
 	Send(portMapping types.PortMapping) error
-}
-
-// VTunnelForwarder forwards the PortMappings to VTunnel Peer process.
-type VTunnelForwarder struct {
-	peerAddr string
-}
-
-func NewVTunnelForwarder(peerAddr string) *VTunnelForwarder {
-	return &VTunnelForwarder{
-		peerAddr: peerAddr,
-	}
-}
-
-// Send forwards the port mappings to Vtunnel Peer.
-func (v *VTunnelForwarder) Send(portMapping types.PortMapping) error {
-	conn, err := net.Dial("tcp", v.peerAddr)
-	if err != nil {
-		return err
-	}
-	defer conn.Close()
-
-	err = json.NewEncoder(conn).Encode(portMapping)
-	if err != nil {
-		return err
-	}
-
-	return nil
 }
